Changed Ds for incident, updated web-gateway to reflect. Updated inmarsat to 313 and exchange to 5.2.12
