+++ unionvms-test/src/test/java/eu/europa/ec/fisheries/uvms/docker/validation/webgateway/dto/ExtendedIncidentLogDto.java
@@ -1,8 +1,5 @@
 package eu.europa.ec.fisheries.uvms.docker.validation.webgateway.dto;
 
-import eu.europa.ec.fisheries.schema.exchange.v1.ExchangeLogStatusType;
-import eu.europa.ec.fisheries.uvms.asset.client.model.Note;
-import eu.europa.ec.fisheries.uvms.docker.validation.movement.model.MicroMovement;
 import eu.europa.ec.fisheries.uvms.incident.model.dto.IncidentLogDto;
 
 import java.util.HashMap;
@@ -10,13 +7,17 @@ import java.util.Map;
 
 public class ExtendedIncidentLogDto {
 
-    private Map<Long, IncidentLogDto> incidentLogs = new HashMap<>();
+    private Map<Long, IncidentLogDto> incidentLogs;
 
-    private Map<String, Note> notes = new HashMap<>();
+    private RelatedObjectDto relatedObjects = new RelatedObjectDto();
 
-    private Map<String, ExchangeLogStatusType> polls = new HashMap<>();
+    public ExtendedIncidentLogDto() {
+        incidentLogs = new HashMap<>();
+    }
 
-    private Map<String, MicroMovement> positions = new HashMap<>();
+    public ExtendedIncidentLogDto(int amountOfLogs) {
+        incidentLogs = new HashMap<>(amountOfLogs);
+    }
 
     public Map<Long, IncidentLogDto> getIncidentLogs() {
         return incidentLogs;
@@ -26,27 +27,11 @@ public class ExtendedIncidentLogDto {
         this.incidentLogs = incidentLogs;
     }
 
-    public Map<String, Note> getNotes() {
-        return notes;
+    public RelatedObjectDto getRelatedObjects() {
+        return relatedObjects;
     }
 
-    public void setNotes(Map<String, Note> notes) {
-        this.notes = notes;
-    }
-
-    public Map<String, ExchangeLogStatusType> getPolls() {
-        return polls;
-    }
-
-    public void setPolls(Map<String, ExchangeLogStatusType> polls) {
-        this.polls = polls;
-    }
-
-    public Map<String, MicroMovement> getPositions() {
-        return positions;
-    }
-
-    public void setPositions(Map<String, MicroMovement> positions) {
-        this.positions = positions;
+    public void setRelatedObjects(RelatedObjectDto relatedObjects) {
+        this.relatedObjects = relatedObjects;
     }
 }
